requirements.txt
audio2numpy==0.1.2
audioread==3.0.1
beautifulsoup4==4.12.2
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
msgpack==1.0.7
mutagen==1.47.0
numba==0.58.1
numpy==1.26.1
packaging==23.2
pandas==2.1.3
platformdirs==3.11.0
pooch==1.8.0
pycparser==2.21
pycryptodomex==3.19.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
https://github.com/python273/vk_api/archive/master.zip
websockets==12.0
youtube-transcript-api==0.6.1
yt-dlp==2023.10.13
tqdm
torchaudio
transformers
auditok==0.2.0
whisper-timestamped==1.15.6
openai-whisper==20240930